codeAnalyser/CodeAnalyserZip/adriraj_code/requirements.txt
colorama
requests
python-dotenv
pygments
langchain-openai
langgraph
PyGithub
streamlit